Vagrant setup enhancements
Using the Gemfile to install the components needed for the installation
in the Vagrant file - get as close to the setup that also occurs in
Travis-CI.

Add a provision command that will run whenever the machine `vagrant
up`'s - the jekyll server will start in the background of the system,
you no longer need to do that manually.

Machine also will now have the htmlproof app installed - same thing that
is being used for the Travis-CI checks, so you can run the same check
that the CI environment does. See instructions in the README.