Vagrantfile mods and add Vagrant-cachier
Start of the Vagrantfile was using an odd first statement - was for an
older version of Vagrant not the newest - updated accordingly.

If anyone has the `vagrant-cachier` plugin installed it will end up
using the local cache it sets up to speed up the download time for
packages when the apt-get installs are running. If you don't have it
nothing will happen.